Add Exception handling for deleting empty Post
해당 post_id에 대한 게시글이 존재하지 않을 경우 redirect하여
exception 처리

JSTL을 이용하여 date format 변경

+++ src/main/java/com/study/controller/cnu/CnuPostController.java
@@ -86,8 +86,13 @@ public class CnuPostController {
         cnuPost.setPostId(postId);
         cnuPost.setIsDel(true);
 
-        String chk_passowrd=cnuRepository.selectCnuPost(postId).getPassword();
+        CnuPost myRepository = cnuRepository.selectCnuPost(postId);
 
+        if(myRepository == null){
+            return "redirect:/post?emptyPost";
+        }
+
+        String chk_passowrd = myRepository.getPassword();
         if(chk_passowrd.equals(password))
         {
             cnuRepository.deleteCnuPost(cnuPost);
@@ -104,8 +109,12 @@ public class CnuPostController {
         cnuPost.setPostId(postId);
         cnuPost.setIsDel(true);
 
+        CnuPost myRepository = cnuRepository.selectCnuPost(postId);
+        if(myRepository == null){
+            return "redirect:/post?emptyPost";
+        }
 
-        String chk_passowrd=cnuRepository.selectCnuPost(postId).getPassword();
+        String chk_passowrd=myRepository.getPassword();
 
         if(chk_passowrd.equals(password))
         {
